+++ src/collective/upgrade/run.py
@@ -29,9 +29,9 @@ parser.add_argument(
     "-u", "--username", help='Specify username to use during the upgrade '
     '(if not provided, a special user will run the upgrade).')
 parser.add_argument(
-    "-p", "--disable-pdb",
+    "-D", "--pdb",
     action="store_true",
-    help='When upgrading a portal disable start of pdb post mortem.'
+    help='When upgrading a portal enable post-mortem debugging.'
 )
 
 group = parser.add_argument_group('upgrades')
@@ -116,10 +116,10 @@ def main(app=None, args=None):
     runner = app.restrictedTraverse('@@collective.upgrade.form')
     try:
         runner.upgrade(**kw)
-    except:
+    except Exception:
         transaction.abort()
         runner.logger.exception('Exception running the upgrades.')
-        if not args.disable_pdb:
+        if args.pdb:
             pdb.post_mortem(sys.exc_info()[2])
         raise